Add dynamic import treeshaking for canvases to avoid double-bundling issue
