+++ src/lib/calc-worker.js
