+++ src/lib/offscreen-worker.js
