+++ src/lib/throttle.js
@@ -1,0 +1,10 @@
+export function throttle(callback, offset) {
+  let baseTime = 0;
+  return (...args) => {
+    const currentTime = Date.now();
+    if (baseTime + offset <= currentTime) {
+      baseTime = currentTime;
+      callback(...args);
+    }
+  };
+}
