+++ src/pages/[slug].astro
@@ -1,6 +1,7 @@
 ---
 import collections from "../data/collections.json";
 import Layout from "../layouts/Layout.astro";
+import CanvasControlFlow from "../components/CanvasControlFlow.svelte";
 
 export async function getStaticPaths() {
   return collections.map((collection) => {
@@ -78,9 +79,7 @@ let { name, slug, length, description } = Astro.props;
     data-barba-namespace="collection"
     class="w-screen h-screen px-16 py-16 bg-[#1c1c1c]"
   >
-    <canvas
-      id="webgl-canvas"
-      class="fixed top-0 left-0 w-screen h-screen overflow-hidden"></canvas>
+    <CanvasControlFlow client:idle />
     <div class="py-1 overflow-hidden">
       <h1
         class="relative uppercase font-sans text-[120px] capsize text-slate-50 mo-text leading-tight mix-blend-difference z-10"
